Fix error case which session change request failed when pausing video
When a user clicks the pause video button, the camera close request and
the session modify request are executed at the same time. However, the
session modify request fails occasionally by a limitation of the Network
side. In that case, the video session is continued with the camera
closed state and a user cannot open the camera again due to pause video
button disabled.

Remove camera close request triggered by the pause video button clicked
because it is handled appropriately according to the result of the
session modify request. Also, enable the pause video button and show
error message to a user if receiving the error result of the session
modify request.

Test: manual
Bug: 69235524

Change-Id: I9a2dde755a6c28edfb0ce962b55ac8a6e907ca97
Signed-off-by: rezaadi0105 <rezaadipangestu5@gmail.com>