Fix issue that the hold label is not shown during remote video holding
Hold label is not shown even if a remote party holds a video call
because VideoCallFragment#updateRemoteOffView is not called. A condition
checking shouldShowRemote whether to call the method was added in

To solve this issue, checks isRemotelyHeld that indicates hold status of
remote party and calls #updateRemoteOffView if it is updated.

Test: manual - Check that the hold label is shown when a remote party
holds a video call
Bug: 110244504

Change-Id: I2dfac18b1a913de50adfbe4cc76f073678f5e186
Signed-off-by: smokey18 <kmboplays@gmail.com>